move float support to its own file
